Remove redundant function
