+++ src/components/analytics/analytics.tsx
@@ -8,9 +8,9 @@ import {
     getLast30DaysUsage,
     getTodayUsage,
     getUsageNumber,
+    formatBytes,
 } from "@lib/analyticsUtils";
 import { Link } from "react-router";
-import StatsCard from "./statsCard";
 import PageLayout from "./pageLayout";
 import useUsage from "@/hooks/useUsage";
 
@@ -21,6 +21,21 @@ function AnalyticsHeader() {
             <span className="font-mono text-sm font-bold tracking-wider uppercase">
                 Usage Analytics for YouTube
             </span>
+        </div>
+    );
+}
+
+function StatsCard({ title, number }: { title: string; number: number }) {
+    const formattedNumber = formatBytes(number);
+    return (
+        <div className="flex flex-col justify-center gap-2.5 rounded-lg border border-neutral-800 bg-neutral-900 py-4 pr-2.5 pl-5.5 hover:cursor-pointer hover:bg-neutral-800">
+            <div className="font-mono text-sm font-semibold text-teal-400 uppercase">{title}</div>
+            <div className="flex justify-between font-mono text-2xl font-bold text-stone-200">
+                {formattedNumber}
+                <div className="flex items-end font-mono text-xs text-teal-600 underline">
+                    View Details →
+                </div>
+            </div>
         </div>
     );
 }
