+++ src/components/analytics/usageDetails.tsx
@@ -1,32 +1,31 @@
-import { getNumVideosWatched, getUsageNumber } from "@lib/analyticsUtils";
+import {
+    getNumVideosWatched,
+    getUsageNumber,
+    getUsageByDate,
+    formatDate,
+} from "@lib/analyticsUtils";
 import { useParams } from "react-router";
 import AnalyticsHeader from "./analyticsHeader";
 import AnalyticsBody from "./analyticsBody";
 import PageLayout from "./pageLayout";
 import useUsage from "@/hooks/useUsage";
 
-function formatDate(date: string) {
-    const dateString = new Date(date).toDateString();
-    const firstSpaceIndex = dateString.indexOf(" ");
-    return dateString.slice(firstSpaceIndex + 1);
-}
-
 export function UsageDetails() {
     const { usage, error } = useUsage();
     const { date } = useParams();
     if (!date) return;
 
-    const todayUsage = usage[date] ?? {};
+    const daysUsage = getUsageByDate(usage, date);
 
     return (
         <PageLayout>
             <AnalyticsHeader
-                numVideosWatched={getNumVideosWatched({ [date]: todayUsage })}
-                title={formatDate(date)}
-                totalDataUsage={getUsageNumber({ [date]: todayUsage })}
+                numVideosWatched={getNumVideosWatched(daysUsage)}
+                title={formatDate(new Date(date))}
+                totalDataUsage={getUsageNumber(daysUsage)}
                 key={date}
             />
-            <AnalyticsBody usage={{ [date]: todayUsage }} error={error} />
+            <AnalyticsBody usage={daysUsage} error={error} />
         </PageLayout>
     );
 }
